refactor(quant): pre-extract UnifiedQuantizationLevel cluster to foundation (Slice D)
Move the quantization-level MODEL cluster down from the modality vector crate to a
new foundation crate `proximadb-quantization-model`, and redirect src/storage's
references to it. This removes the divergent-types trap (foundation's simple
QuantizationLevel ≠ modality's rich one) that blocked both the storage-engine port
and clean engine extraction (Slice E).

The cluster (UnifiedQuantizationLevel + the rich QuantizationLevel enum +
Product/Scalar/Binary/Uniform/No/Custom/TurboQuantization variant payloads + their
QuantizationMethod impls) was canonically in one place
(proximadb-vector/src/quantization/internal_types.rs) and is foundation-pure (only
deps: proximadb-quantization-types + serde). Moved verbatim (~620 lines) to the new
foundation crate; internal_types.rs is now a `pub use proximadb_quantization_model::*`
re-export (so the modality/kernel/compute::quantization chain is unchanged).

B2 — redirect src/storage's ~43 cluster-type refs
(crate::compute::quantization::{quantization_engine,types}::{UnifiedQuantizationLevel,
QuantizationLevel,ProductQuantization}) to proximadb_quantization_model::* (Mechanism A:
pure path rewrite, downward dep). Behavior-identical (same types, new location).

- crates/foundation/proximadb-quantization-model: new crate (the cluster).
- proximadb-vector: re-export + dep + experimental-turboquant feature propagation.
- root Cargo.toml: members + workspace deps + turboquant aggregator + root dep.
- src/storage: ~13 engine/format files redirected.

Verified: cargo check --lib + clippy --lib clean; layering check green; storage up-edge
ratchet 248 -> 205 (-43). Behavior-identical type move (recall-neutral — no format/logic
change). Unblocks the QuantizationEnginePort adoption + Slice E engine extraction.

Per ROOT_CRATE_DECOMPOSITION_PLAN + SLICE_D (the divergent-types trap).
