+++ crates/foundation/proximadb-quantization-model/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "proximadb-quantization-model"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Canonical quantization-level model for ProximaDB: UnifiedQuantizationLevel + the rich QuantizationLevel enum + variant payloads (foundation-tier, shared by storage and the modality)"
+
+[features]
+# Gates the TurboQuant variant/payload + calibration paths (moved verbatim from
+# proximadb-vector/src/quantization/internal_types.rs). Default off; same gate as
+# proximadb-quantization-types' experimental-turboquant.
+experimental-turboquant = []
+
+[dependencies]
+proximadb-quantization-types = { workspace = true }
+serde = { workspace = true }
+
+[lints]
+workspace = true
