+++ src/storage/engines/core/formats/proximablocks/block_structures.rs
@@ -266,8 +266,7 @@ pub struct ProximaDataBlock {
     /// Quantized vectors using unified engine
     pub quantized_vectors: Option<Vec<Vec<u8>>>,
     /// Quantization level used
-    pub quantization_level:
-        Option<crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel>,
+    pub quantization_level: Option<proximadb_quantization_model::UnifiedQuantizationLevel>,
 
     /// SIMD-encoded vector data (when layout != FullVector)
     /// Stores transposed and encoded dimensions for SIMD operations
