+++ src/storage/engines/sst/compactor_impl.rs
@@ -1296,17 +1296,14 @@ mod tests {
 
         let storage_config =
             crate::compute::quantization::storage_engine::StorageQuantizationConfig {
-                primary_level: Some(
-                    crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel::pq8(32),
-                ),
+                primary_level: Some(proximadb_quantization_model::UnifiedQuantizationLevel::pq8(
+                    32,
+                )),
                 filter_level: Some(
-                    crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel::binary(),
+                    proximadb_quantization_model::UnifiedQuantizationLevel::binary(),
                 ),
-                fast_level: Some(
-                    crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel::int8(),
-                ),
-                distance_metric:
-                    proximadb_distance_kernel::engine::DistanceMetric::Cosine,
+                fast_level: Some(proximadb_quantization_model::UnifiedQuantizationLevel::int8()),
+                distance_metric: proximadb_distance_kernel::engine::DistanceMetric::Cosine,
                 enable_progressive: true,
                 filter_threshold: 100.0,
                 candidate_multiplier: 4,
