+++ src/storage/engines/sst/readers/sst_query_engine.rs
@@ -883,7 +883,7 @@ impl ModularBlockReader {
     ) -> Result<
         Option<(
             Vec<Vec<u8>>,
-            crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel,
+            proximadb_quantization_model::UnifiedQuantizationLevel,
         )>,
     > {
         // For now, read the entire block and extract quantized section
