+++ src/storage/engines/swift/engine.rs
@@ -406,17 +406,14 @@ impl SwiftEngine {
         // Configure storage quantization for SWIFT (SST-based engine)
         let storage_config =
             crate::compute::quantization::storage_engine::StorageQuantizationConfig {
-                primary_level: Some(
-                    crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel::pq8(16),
-                ),
+                primary_level: Some(proximadb_quantization_model::UnifiedQuantizationLevel::pq8(
+                    16,
+                )),
                 filter_level: Some(
-                    crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel::binary(),
+                    proximadb_quantization_model::UnifiedQuantizationLevel::binary(),
                 ),
-                fast_level: Some(
-                    crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel::int8(),
-                ),
-                distance_metric:
-                    proximadb_distance_kernel::engine::DistanceMetric::Cosine,
+                fast_level: Some(proximadb_quantization_model::UnifiedQuantizationLevel::int8()),
+                distance_metric: proximadb_distance_kernel::engine::DistanceMetric::Cosine,
                 enable_progressive: true,
                 filter_threshold: 100.0,
                 candidate_multiplier: 10,
