+++ src/storage/engines/viper/pipeline.rs
@@ -102,8 +102,7 @@ pub struct ProcessingConfig {
     pub sorting_strategy: SortingStrategy,
 
     /// Vector quantization level (None = no quantization)
-    pub quantization_level:
-        Option<crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel>,
+    pub quantization_level: Option<proximadb_quantization_model::UnifiedQuantizationLevel>,
 }
 
 /// Flushing configuration
@@ -1080,7 +1079,7 @@ impl VectorRecordProcessor {
     async fn apply_vector_quantization(
         &self,
         records: &[VectorRecord],
-        quantization_level: crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel,
+        quantization_level: proximadb_quantization_model::UnifiedQuantizationLevel,
     ) -> Result<Vec<crate::compute::quantization::quantization_engine::QuantizedVector>> {
         if records.is_empty() {
             return Ok(vec![]);
@@ -2201,7 +2200,7 @@ impl ParquetFlusher {
 
         // Group quantized vectors by quantization level for efficient column storage
         let mut quantization_groups: std::collections::HashMap<
-            crate::compute::quantization::quantization_engine::UnifiedQuantizationLevel,
+            proximadb_quantization_model::UnifiedQuantizationLevel,
             Vec<&crate::compute::quantization::quantization_engine::QuantizedVector>,
         > = std::collections::HashMap::new();
 
@@ -2215,7 +2214,7 @@ impl ParquetFlusher {
         // Create columns for each quantization level
         for (level, qvecs) in quantization_groups {
             match level.level_type {
-                Some(crate::compute::quantization::types::QuantizationLevel::Pq(pq)) => {
+                Some(proximadb_quantization_model::QuantizationLevel::Pq(pq)) => {
                     let bits_per_code = pq.bits_per_code;
                     let num_subvectors = pq.num_subvectors;
                     // Store PQ codes as binary data
@@ -2241,7 +2240,7 @@ impl ParquetFlusher {
                     }
                 }
 
-                Some(crate::compute::quantization::types::QuantizationLevel::Scalar(scalar)) => {
+                Some(proximadb_quantization_model::QuantizationLevel::Scalar(scalar)) => {
                     let bits = scalar.bits;
                     // Store uniform quantization as packed binary data
                     let mut data_owned: Vec<Vec<u8>> = Vec::new(); // All owned data
@@ -2276,24 +2275,24 @@ impl ParquetFlusher {
                     continue;
                 }
 
-                Some(crate::compute::quantization::types::QuantizationLevel::None(_)) => {
+                Some(proximadb_quantization_model::QuantizationLevel::None(_)) => {
                     // No quantization - skip
                     continue;
                 }
 
-                Some(crate::compute::quantization::types::QuantizationLevel::Uniform(_)) => {
+                Some(proximadb_quantization_model::QuantizationLevel::Uniform(_)) => {
                     // Uniform quantization - handle similar to INT8
                     // Deferred: Implement uniform quantization handling
                     continue;
                 }
 
-                Some(crate::compute::quantization::types::QuantizationLevel::Binary(_)) => {
+                Some(proximadb_quantization_model::QuantizationLevel::Binary(_)) => {
                     // Binary quantization - handle as binary data
                     // Deferred: Implement binary quantization handling
                     continue;
                 }
 
-                Some(crate::compute::quantization::types::QuantizationLevel::Custom(custom)) => {
+                Some(proximadb_quantization_model::QuantizationLevel::Custom(custom)) => {
                     let bits_per_element = custom.bits_per_element;
                     // Handle custom quantization similar to uniform
                     let mut data_owned: Vec<Vec<u8>> = Vec::new(); // All owned data
@@ -2326,7 +2325,7 @@ impl ParquetFlusher {
                 // pipeline below — they have their own `.tq` wire format
                 // (LLD §3). Skip here; P8 wires the durable storage path.
                 #[cfg(feature = "experimental-turboquant")]
-                Some(crate::compute::quantization::types::QuantizationLevel::TurboQuant(_)) => {
+                Some(proximadb_quantization_model::QuantizationLevel::TurboQuant(_)) => {
                     debug!(
                         "TurboQuant quantization level encountered in VIPER pipeline; \
                          skipping (storage routes through .tq files per TURBOQUANT_LLD §3)",
@@ -2342,7 +2341,7 @@ impl ParquetFlusher {
 
         for qvec in quantized_vectors {
             // Get bits per value based on quantization level
-            use crate::compute::quantization::types::QuantizationLevel;
+            use proximadb_quantization_model::QuantizationLevel;
             let (bits, is_product_quant) = match &qvec.quantization_level.level_type {
                 Some(QuantizationLevel::Binary(_)) => (1, false),
                 Some(QuantizationLevel::Scalar(s)) => (s.bits as u8, false),
@@ -3655,7 +3654,7 @@ impl Default for ViperPipelineConfig {
                 compression: true,
                 sorting_strategy: SortingStrategy::ByTimestamp,
                 quantization_level: Some(
-                    crate::compute::quantization::types::UnifiedQuantizationLevel::Int8,
+                    proximadb_quantization_model::UnifiedQuantizationLevel::Int8,
                 ), // Default to 8-bit quantization
             },
             flushing_config: FlushingConfig {
